Rename EntityLayout component into EntityContainerWithFilterBar
