+++ mathesar_ui/src/components/EntityContainerWithFilterBar.svelte
