+++ mathesar_ui/src/pages/schema/SchemaTables.svelte
@@ -2,8 +2,8 @@
   import type { TableEntry } from '@mathesar/api/types/tables';
   import type { Database, SchemaEntry } from '@mathesar/AppTypes';
   import { labeledCount } from '@mathesar/utils/languageUtils';
+  import EntityContainerWithFilterBar from '@mathesar/components/EntityContainerWithFilterBar.svelte';
   import TablesList from './TablesList.svelte';
-  import EntityLayout from '../../components/EntityLayout.svelte';
   import CreateNewTableTutorial from './CreateNewTableTutorial.svelte';
   import CreateNewTableButton from './CreateNewTableButton.svelte';
 
@@ -33,7 +33,7 @@
   }
 </script>
 
-<EntityLayout
+<EntityContainerWithFilterBar
   searchPlaceholder="Search Tables"
   bind:searchQuery={tableSearchQuery}
   on:clear={clearQuery}
@@ -57,4 +57,4 @@
       <TablesList {canExecuteDDL} tables={filteredTables} {database} {schema} />
     {/if}
   </svelte:fragment>
-</EntityLayout>
+</EntityContainerWithFilterBar>
